pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
license-expression = "MIT"
dependencies = [
    "django>=4.0.3",
    "dj-database-url>=0.5.0",
    "psycopg2-binary>=2.9.3",
    "django-encrypted-model-fields>=0.6.5",
    "boto3>=1.21.30",
    "social-auth-app-django>=5.0.0",
    "redis>=4.2.1",
    "gunicorn>=20.1.0",
    "uvicorn[standard]>=0.17.6",
    "whitenoise>=6.0.0",
    "huey>=2.4.3",
]
requires-python = ">=3.9"

[project.urls]
Homepage = ""
[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
_.env_file = ".env"
req = "pdm export -o requirements.txt --without-hashes --prod"
